Updated the Cloud Function to delete the summary node for a month when no detail nodes exist. Changed category limit to be a double. Added checks each time a Map is iterated to make sure the Map is not null. Changed the BaseActivity variable name for the Group HashMap. Fixed monthly limit lookup to use the month that corresponds to the receipt date. Updated join/leave group to create/delete a category for the user.

+++ mvlab7/app/src/main/java/com/example/myapplication/Category.java
@@ -8,11 +8,11 @@ public class Category {
 
     public String displayName;
 
-    public Integer limit;  // -1 is unlimited
+    public Double limit;  // -1.0 is unlimited
 
     public Category() {}
 
-    public Category(String name, String displayName, Integer limit) {
+    public Category(String name, String displayName, Double limit) {
         this.name = name;
         this.displayName = displayName;
         this.limit = limit;
